backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Upload
aiofiles==23.2.1

# Background Tasks
APScheduler==3.10.4

# Email
fastapi-mail==1.4.1

# Date/Time
python-dateutil==2.8.2

# CORS
starlette==0.35.1

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
